+++ lib/src/job/launcher/parallel_launcher.dart
@@ -11,6 +11,7 @@ import 'package:async_task/async_task.dart';
 // Project imports:
 import 'package:batch/batch.dart';
 import 'package:batch/src/job/launcher/launcher.dart';
+import 'package:batch/src/job/parallel/parallel_executor.dart';
 
 class ParallelLauncher extends Launcher<Parallel> {
   /// Returns the new instance of [ParallelLauncher].
@@ -27,9 +28,11 @@ class ParallelLauncher extends Launcher<Parallel> {
   Future<void> run() async => await super.executeRecursively(
         event: _parallel,
         execute: (parallel) async {
+          final executors = _buildExecutor(parallel.tasks);
+
           final asyncExecutor = AsyncExecutor(
-            parallelism: parallel.executors.length,
-            taskTypeRegister: () => parallel.executors,
+            parallelism: executors.length,
+            taskTypeRegister: () => executors,
             logger: (String type, dynamic message,
                     [dynamic error, dynamic stackTrace]) =>
                 info(message),
@@ -38,10 +41,10 @@ class ParallelLauncher extends Launcher<Parallel> {
           asyncExecutor.logger.enabled = true;
 
           try {
-            final executions = asyncExecutor.executeAll(parallel.executors);
+            final executions = asyncExecutor.executeAll(executors);
             await Future.wait(executions);
 
-            for (final executor in parallel.executors) {
+            for (final executor in executors) {
               for (final isolatedMessage in executor.result!) {
                 isolatedMessage.output();
               }
@@ -51,4 +54,18 @@ class ParallelLauncher extends Launcher<Parallel> {
           }
         },
       );
+
+  List<ParallelExecutor> _buildExecutor(final List<ParallelTask> tasks) {
+    final executors = <ParallelExecutor>[];
+    for (final task in tasks) {
+      executors.add(
+        ParallelExecutor(
+          parallelTask: task,
+          context: context,
+        ),
+      );
+    }
+
+    return executors;
+  }
 }