+++ lib/src/log/logger.dart
@@ -20,9 +20,8 @@ import 'package:batch/src/log/printer/log_printer.dart';
 
 class Logger {
   /// Returns the new instance of [Logger].
-  Logger.loadFrom({
-    required LogConfiguration config,
-  })  : _filter = config.filter ?? DefaultLogFilter(),
+  Logger.loadFrom({required LogConfiguration config})
+      : _filter = config.filter ?? DefaultLogFilter(),
         _printer = config.printer ?? DefaultLogPrinter(),
         _output = config.output ?? ConsoleLogOutput(),
         _printLog = config.printLog {